sorting  code to files and init ndr field export
